SWM-26 refactor: enums 분리 및 디자인 수정
