+++ app/src/main/java/com/betterlife/antifragile/data/model/enums/Gender.kt
@@ -1,0 +1,6 @@
+package com.betterlife.antifragile.data.model.enums
+
+enum class Gender(val toKorean: String) {
+    MALE("남성"),
+    FEMALE("여성")
+}
