+++ app/src/main/java/com/betterlife/antifragile/data/model/enums/LoginType.kt
@@ -1,0 +1,7 @@
+package com.betterlife.antifragile.data.model.enums
+
+enum class LoginType(val toKorean: String) {
+    GOOGLE("구글"),
+    NAVER("네이버"),
+    KAKAO("카카오")
+}
